templates.js revamp
