+++ express/blocks/template-list/template-list.css
@@ -700,7 +700,6 @@ main .template-list.fullwidth.sm-view > .masonry-col {
   main .template-list-fullwidth-apipowered-container > div,
   main .template-list-fullwidth-apipowered-container > .default-content-wrapper {
     max-width: none;
-    padding: 0 28px;
   }
 
   main .template-list.sixcols > .masonry-col,
