+++ express/scripts/scripts.js
@@ -2323,8 +2323,12 @@ async function loadEager() {
   }
   if (!window.hlx.lighthouse) await decorateTesting();
 
-  if (window.location.href.includes('/express/templates/')) {
-    await import('./templates.js');
+  if (getMetadata('sheet-powered') === 'Y') {
+    import('./content-replace.js');
+  }
+
+  if (getMetadata('template-search-page') === 'Y') {
+    await import('./template-redirect.js');
   }
 
   if (main) {
@@ -2337,7 +2341,7 @@ async function loadEager() {
     displayOldLinkWarning();
     wordBreakJapanese();
 
-    const lcpBlocks = ['columns', 'hero-animation', 'hero-3d', 'template-list', 'floating-button', 'fullscreen-marquee', 'collapsible-card'];
+    const lcpBlocks = ['columns', 'hero-animation', 'hero-3d', 'template-list', 'floating-button', 'fullscreen-marquee', 'collapsible-card', 'search-marquee'];
     const block = document.querySelector('.block');
     const hasLCPBlock = (block && lcpBlocks.includes(block.getAttribute('data-block-name')));
     if (hasLCPBlock) await loadBlock(block, true);
@@ -2436,6 +2440,9 @@ export async function addFreePlanWidget(elem) {
 async function loadLazy() {
   const main = document.querySelector('main');
 
+  if (window.location.href.includes('/express/templates/')) {
+    await import('./templates.js');
+  }
   // post LCP actions go here
   sampleRUM('lcp');
